+++ lib/portrayal/schema.rb
@@ -54,6 +54,11 @@ module Portrayal
       def eql?(other); self.class == other.class && self == other end
       def hash; [self.class, self.class.portrayal.attributes(self)].hash end
 
+      def freeze
+        self.class.portrayal.attributes(self).values.each(&:freeze)
+        super
+      end
+
       def ==(other)
         self.class.portrayal.attributes(self) ==
           self.class.portrayal.attributes(other)